렌더타겟 설정 및 렌더 스텐실 설정은 매 프레임마다 해주는 것으로 변경

+++ HoFramework/Define.h
@@ -83,7 +83,7 @@ struct Mesh
 {
 public:
 	vector<Vertex> vertices;
-	vector<uint16_t> indices;
+	vector<uint32_t> indices;
 	string textureSourceName;
 };
